Move make_api_calls to its own module at crates/fuzzing/src/oracles/api.rs
